+++ resources/todomvc/vanilla-examples/javascript-es6-webpack/src/app.js
@@ -1,8 +1,36 @@
+import View from "./view";
+import Controller from "./controller";
+import Model from "./model";
+import Store from "./store";
+import Template from "./template";
+
 import "todomvc-app-css/index.css";
 import "./app.css";
 
-import { updateTodo } from "./todo";
+let todo;
+const onChange = () => {
+    todo.controller.setView(document.location.hash);
+};
 
-export function onLoad() {
-    updateTodo();
+const onLoad = () => {
+    todo = new Todo("javascript-es6-webpack");
+    onChange();
+};
+
+function Todo(name) {
+    this.storage = new Store(name);
+    this.model = new Model(this.storage);
+    this.template = new Template();
+    this.view = new View(this.template);
+    this.controller = new Controller(this.model, this.view);
 }
+
+if (module.hot) {
+    module.hot.accept(function (err) {});
+    if (document.readyState === "complete") {
+        onLoad();
+    }
+}
+
+window.addEventListener("load", onLoad);
+window.addEventListener("hashchange", onChange);
