+++ resources/todomvc/vanilla-examples/javascript-es6-webpack/webpack.common.js
@@ -2,7 +2,7 @@ const path = require("path");
 
 module.exports = {
     entry: {
-        app: "./src/index.js",
+        app: "./src/app.js",
     },
     output: {
         filename: "[name].bundle.js",
